+++ src/main/java/com/futurenbetter/saas/modules/promotion/entity/PromotionUsage.java
@@ -24,7 +24,7 @@ public class PromotionUsage {
     private Long promotionUsageId;
 
     @Column(name = "discount_amount")
-    private Double discountAmount;
+    private Float discountAmount;
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
